+++ docx4j-export-fo-tests/src/test/java/org/docx4j/convert/out/XSLFO/PStyle12PtInTableNormalOverrideTrueTest.java
@@ -110,7 +110,7 @@ public class PStyle12PtInTableNormalOverrideTrueTest extends PStyleTableAbstract
 
 		setSetting(wordMLPackage, OVERRIDE);  // table style should get overridden
 		
-		wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+		wordMLPackage.save(outputDocx());
 		
 		// NB PropertyResolver puts 10pt in DocDefaults, if nothing is specified!
 		
@@ -142,13 +142,13 @@ public class PStyle12PtInTableNormalOverrideTrueTest extends PStyleTableAbstract
 		
 		setSetting(wordMLPackage, OVERRIDE);  // table style should get overridden
 	
-		wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+		wordMLPackage.save(outputDocx());
 		
 		ParagraphStylesInTableFix.process(wordMLPackage);
 		
 //		// Revert style and save: 
 //		ppr.setPStyle(ps); // doesn't work - wrong ref!
-//		wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+//		wordMLPackage.save(outputDocx());
 		
 		Style ours = null;
 		for (Style s : wordMLPackage.getMainDocumentPart().getStyleDefinitionsPart().getContents().getStyle()) {
@@ -184,13 +184,13 @@ public class PStyle12PtInTableNormalOverrideTrueTest extends PStyleTableAbstract
 		
 		setSetting(wordMLPackage, OVERRIDE); 
 
-		wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+		wordMLPackage.save(outputDocx());
 		
 		ParagraphStylesInTableFix.process(wordMLPackage);
 		
 //		// Revert style and save: 
 //		ppr.setPStyle(ps); // doesn't work - wrong ref!
-//		wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+//		wordMLPackage.save(outputDocx());
 		
 		Style ours = null;
 		for (Style s : wordMLPackage.getMainDocumentPart().getStyleDefinitionsPart().getContents().getStyle()) {
@@ -211,7 +211,7 @@ public class PStyle12PtInTableNormalOverrideTrueTest extends PStyleTableAbstract
 		
 		WordprocessingMLPackage wordMLPackage = test(mdpXml_direct_12pt, styles_no_font_sz, 20); // uses implicit DocDefault, but irrelevant
 		
-		//wordMLPackage.save(new File(System.getProperty("user.dir") + "/OUT_PStyleInTableTest.docx"));
+		//wordMLPackage.save(outputDocx());
 		
 		/* In this case, our result correctly preserves the direct rPr formatting
 		 * (so the contents of Normal-TableNormal-BR is irrelevant)